refactor(creative_agent): extract static gallery HTML to gallery_template.py
